+++ engine/src/semulator/impl/api/basic/OpDecrease.java
@@ -12,10 +12,10 @@ import java.util.List;
 public class OpDecrease extends AbstractOpBasic {
     @Override
     public Label execute(SProgram program) {
-        Long variableValue = program.getVariableValue(getVariable());
+        Integer variableValue = program.getVariableValue(getVariable());
         variableValue = Math.max(0, variableValue - 1);
         ArrayList<Variable> vars = new ArrayList<>();
-        ArrayList<Long> vals = new ArrayList<>();
+        ArrayList<Integer> vals = new ArrayList<>();
         vars.add(getVariable());
         vals.add(variableValue);
